% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CancerGram_model}
\alias{CancerGram_model}
\title{Data and model for CancerGram}
\format{
A list of length three: random forest model for mers, random forest
model for peptides and informative n-grams used to make predictions.
}
\description{
Random forest models grown using the \code{ranger} package 
and informative n-grams
}
\seealso{
\code{\link[ranger]{ranger}}
}
\keyword{datasets}
