% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerGramModel-package.R
\docType{package}
\name{CancerGramModel-package}
\alias{CancerGramModel-package}
\alias{CancerGramModel}
\title{Data and model storage for CancerGram}
\description{
This package is used only as model and data storage for 
[CancerGram](https://github.com/ksidorczuk/CancerGram) - package for 
prediction of anticancer peptides using n-gram encoding and random forests.
}
\details{
The **CancerGram_model.rda** file stores the core functionality of CancerGram:
the stacked random forest model and list of informative n-grams used for 
classification of peptides/proteins as ACPs, AMPs or Negative. It is essential 
for proper use of CancerGram but needs to be stored in the external repository 
due to large size of the model and CRAN file size limit.
}
\keyword{package}
